+++ CHANGELOG.md
@@ -7,6 +7,20 @@
 ### 🐞 Bug fixes
 - _...Add new stuff here..._
 
+## 24.1.0
+
+### ✨ Features and improvements
+
+- Added `global-state` parameter to `featureFilter` function ([#1279](https://github.com/maplibre/maplibre-style-spec/pull/1279))
+
+## 24.0.0
+
+### ✨ Features and improvements
+
+- ⚠️ Change passing `global-state` to expression from the `evaluate` function to the constructor ([#1267](https://github.com/maplibre/maplibre-style-spec/pull/1267)) - you can add an empty object for global state to overcome this.
+
+### 🐞 Bug fixes
+
 ## 23.3.0
 
 ### ✨ Features and improvements